requirements.txt
Django==3.2.6
google-auth==2.0.2
google-api-python-client==2.11.0
